Refactor JavaScript files
Hopefully, this serves better maintainability in the future.
