+++ _assets/js/github_widget.js
@@ -1,0 +1,51 @@
+//= require 'vendor/jquery-2.1.4.min'
+//= require 'vendor/jquery.githubRepoWidget'
+//= require 'vendor/jXHR'
+
+var github = (function () {
+    function escapeHtml(str) {
+        return $('<div/>').text(str).html();
+    }
+
+    function render(target, repos) {
+        var i = 0, fragment = '', t = $(target)[0];
+
+        for (i = 0; i < repos.length; i++) {
+            fragment += '<a class="list-group-item" href="' + repos[i].html_url + '">';
+            fragment += '<h4 class="list-group-item-heading">' + repos[i].name;
+            if (repos[i].language) {
+                fragment += '<span class="badge pull-right">' + escapeHtml(repos[i].language) + '</span>'
+            }
+            fragment += '</h4><p>' + escapeHtml(repos[i].description || '') + '</p></a>';
+        }
+        t.innerHTML = fragment;
+    }
+
+    return {
+        showRepos: function (options) {
+            $.ajax({
+                url: "https://api.github.com/users/" + options.user + "/repos?sort=pushed&callback=?",
+                dataType: 'jsonp',
+                error: function (err) {
+                    $(options.target + ' li.loading').addClass('error').text("Error loading feed");
+                },
+                success: function (data) {
+                    var repos = [];
+                    if (!data || !data.data) {
+                        return;
+                    }
+                    for (var i = 0; i < data.data.length; i++) {
+                        if (options.skip_forks && data.data[i].fork) {
+                            continue;
+                        }
+                        repos.push(data.data[i]);
+                    }
+                    if (options.count) {
+                        repos.splice(options.count);
+                    }
+                    render(options.target, repos);
+                }
+            });
+        }
+    };
+})();
