+++ _includes/after_footer.html
@@ -1,2 +1,2 @@
 <!-- Latest compiled and minified JavaScript -->
-{% js custom %}
+{% js main %}
